- add patches.drivers/drm-cvs-20050705-update   Update DRM drivers to latest CVS version. - add patches.fixes/gcc4-correct-signed-unsigned-mismatch   Fix signed/unsigned mismatches from gcc-4 - Update configs.
suse-commit: b490fb7420d8f925772ec1f2b530c31c784c0aab

+++ drivers/char/drm/via_ds.h
@@ -30,7 +30,7 @@
 
 /* Set Data Structure */
 #define SET_SIZE 5000
-typedef unsigned int ITEM_TYPE;
+typedef unsigned long ITEM_TYPE;
 
 typedef struct {
 	ITEM_TYPE val;
